requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
crispy-bootstrap5==0.7
    # via -r requirements.in
cryptography==38.0.3
    # via
    #   django-rest-knox
    #   pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-database-url==1.0.0
    # via environs
dj-email-url==1.0.6
    # via environs
django==4.1.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   dj-database-url
    #   django-allauth
    #   django-cors-headers
    #   django-rest-knox
    #   djangorestframework
django-allauth==0.51.0
    # via -r requirements.in
django-cache-url==3.4.2
    # via environs
django-cors-headers==3.13.0
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-rest-knox==4.2.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   django-rest-knox
environs[django]==9.5.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
marshmallow==3.18.0
    # via environs
oauthlib==3.2.2
    # via requests-oauthlib
packaging==21.3
    # via marshmallow
psycopg2-binary==2.9.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via django-allauth
pyparsing==3.0.9
    # via packaging
python-dotenv==0.21.0
    # via environs
python3-openid==3.2.0
    # via django-allauth
pytz==2022.6
    # via djangorestframework
requests==2.28.1
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sqlparse==0.4.3
    # via django
tzdata==2022.6
    # via django
urllib3==1.26.12
    # via requests
whitenoise==6.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
